arturito_microservice/requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.2
boto3==1.10.8
botocore==1.13.8
Click==7.0
cycler==0.10.0
docutils==0.15.2
Flask==1.1.1
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.3
gunicorn==19.9.0
h5py==2.10.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mne==0.19.1
numpy==1.17.3
Pillow==6.2.1
protobuf==3.10.0
pyEDFlib==0.1.15
pylint==2.4.3
pymongo==3.9.0
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
rope==0.14.0
s3transfer==0.2.1
scipy==1.3.1
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typed-ast==1.4.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
secure-smtplib